Code/DL/MCP/quickstart/pyproject.toml
[project]
name = "quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx[socks]>=0.28.1",
    "mcp[cli]>=1.13.1",
    "openai>=1.102.0",
    "pillow>=11.3.0",
]
